Major code cleanup
