+++ include/bytecode_vm/decl_vm.hpp
@@ -11,8 +11,8 @@
 #pragma once
 
 #include "utils/algorithm.hpp"
+#include "utils/files.hpp"
 #include "utils/misc.hpp"
-#include "utils/serialization.hpp"
 
 #include <array>
 #include <concepts>
@@ -141,7 +141,7 @@ public:
   };
 
   template <auto I> static std::tuple_element_t<I, attribute_tuple_type> decode_attribute(auto &first, auto last) {
-    auto [val, iter] = paracl::utils::read_little_endian<std::tuple_element_t<I, attribute_tuple_type>>(first, last);
+    auto [val, iter] = ::utils::read_little_endian<std::tuple_element_t<I, attribute_tuple_type>>(first, last);
     if (!val) throw vm_error{"Decoding error"};
     first = iter;
     return val.value();
@@ -250,7 +250,7 @@ public:
     auto current_instruction = instruction_set.instruction_lookup_table[*(m_execution_context.m_ip++)];
 
     // clang-format off
-    std::visit(paracl::utils::visitors{
+    std::visit(::utils::visitors{
       [this](std::monostate) {
         m_execution_context.halt();
         throw vm_error{"Unknown opcode"};},
