+++ include/frontend/source.hpp
@@ -10,7 +10,7 @@
 
 #pragma once
 
-#include "utils/serialization.hpp"
+#include "utils/files.hpp"
 
 #include <algorithm>
 #include <cassert>
