+++ src/pclc.cc
@@ -22,8 +22,6 @@
 #include <stdexcept>
 #include <string>
 
-namespace utils = paracl::utils;
-
 int main(int argc, char *argv[]) try {
   std::string input_file_name;
   bool dump_binary = false;
